Comments Thomas 11/6/2013
Prabha,

Thanks for the great account of what GeoNode is actually. Let me comment shortly on your last paragraph/sentence. Have you conducted an investigation of why it was so painful to install ? Did you have a chance to document your issues in order to help other joiners avoid the problems you had ?